kontrol/{state_record,prove}: move more helpers for state recording to state_record module
